move rf supplies to supplies sub config with cup dependency
