+++ addons/supplies/rf/CfgVehicles.hpp
