+++ addons/supplies/rf/script_component.hpp
@@ -1,0 +1,4 @@
+#define COMPONENT supplies_rf
+#define COMPONENT_BEAUTIFIED Supplies RF
+#include "\x\tacgt\addons\main\script_mod.hpp"
+#include "\x\tacgt\addons\main\script_macros.hpp"
